hangman version 3
If the user guesses a correct letter, the letter now replaces the blank line. If the user guesses a letter incorrectly, the terminal will then tell the user that the letter is not in the word and will also tell the user how many guesses are left before the end of the game.